requirements.txt
google-generativeai
PyMuPDF
python-dotenv
Flask
Flask-CORS
gunicorn